+++ stubs/inertia/resources/js/Components/PrimaryButton.vue
