+++ stubs/inertia/resources/js/Pages/Teams/Partials/CreateTeamForm.vue
@@ -1,6 +1,6 @@
 <script setup>
 import { useForm } from '@inertiajs/inertia-vue3';
-import Button from '@/Components/Button.vue';
+import PrimaryButton from '@/Components/PrimaryButton.vue';
 import FormSection from '@/Components/FormSection.vue';
 import Input from '@/Components/Input.vue';
 import InputError from '@/Components/InputError.vue';
@@ -58,9 +58,9 @@ const createTeam = () => {
         </template>
 
         <template #actions>
-            <Button :class="{ 'opacity-25': form.processing }" :disabled="form.processing">
+            <PrimaryButton :class="{ 'opacity-25': form.processing }" :disabled="form.processing">
                 Create
-            </Button>
+            </PrimaryButton>
         </template>
     </FormSection>
 </template>
