+++ stubs/inertia/resources/js/Pages/Teams/Partials/UpdateTeamNameForm.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { useForm } from '@inertiajs/inertia-vue3';
 import ActionMessage from '@/Components/ActionMessage.vue';
-import Button from '@/Components/Button.vue';
+import PrimaryButton from '@/Components/PrimaryButton.vue';
 import FormSection from '@/Components/FormSection.vue';
 import Input from '@/Components/Input.vue';
 import InputError from '@/Components/InputError.vue';
@@ -72,9 +72,9 @@ const updateTeamName = () => {
                 Saved.
             </ActionMessage>
 
-            <Button :class="{ 'opacity-25': form.processing }" :disabled="form.processing">
+            <PrimaryButton :class="{ 'opacity-25': form.processing }" :disabled="form.processing">
                 Save
-            </Button>
+            </PrimaryButton>
         </template>
     </FormSection>
 </template>
